Include compiled.h instead of src/compiled.h
This is beneficial for versions of GAP which are installed systemwide
(via Debian, Ubuntu, Fedora, ... package managers), which currently have
to create a fake `src` dir just for the sake of packages like this.

This requires GAP >= 4.9, so also require that. Since no tests are done
with older GAP versions, that seem like a good idea anyway?